pyproject.toml
[project]
name = "nopasanada"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "flask>=3.1.0",
    "fpdf>=1.7.2",
    "img2pdf>=0.6.1",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pyopenssl>=25.1.0",
    "pypdfium2>=4.30.1",
    "requests>=2.32.3",
    "schedule>=1.2.2",
    "selenium>=4.32.0",
    "speechrecognition>=3.14.2",
]
